Tou-01: renamed api.js file to ApiRequest
